Add files via upload
its honestly a mess of files to look at, so if anyone is looking at this then just know theyre for me to look at as a cheat sheet and not to look pretty. oh yeah and no styling cause I was learning to add mysql and learning more react so.